Move widgets to its own dir
