+++ lib/screens/discover.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:http/http.dart' as http;
 
-import '../common_widgets.dart';
+import '../widgets/common_widgets.dart';
 import '../viewmodels.dart';
 import 'newtwt.dart';
