+++ lib/widgets/common_widgets.dart
@@ -9,15 +9,14 @@ import 'package:provider/provider.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'package:goryon/screens/profile.dart';
-import 'package:video_player/video_player.dart';
 
-import 'api.dart';
-import 'models.dart';
-import 'screens/discover.dart';
-import 'screens/follow.dart';
-import 'screens/newtwt.dart';
-import 'screens/timeline.dart';
-import 'viewmodels.dart';
+import '../api.dart';
+import '../models.dart';
+import '../screens/discover.dart';
+import '../screens/follow.dart';
+import '../screens/newtwt.dart';
+import '../screens/timeline.dart';
+import '../viewmodels.dart';
 import 'package:path/path.dart' as path;
 
 class Avatar extends StatelessWidget {
@@ -249,7 +248,7 @@ class _PostListState extends State<PostList> {
               (_, idx) {
                 final twt = widget.twts[idx];
 
-                Function onTap = user.getNickFromTwtxtURL(
+                Function onTwterTap = user.getNickFromTwtxtURL(
                           user.profile.uri.toString(),
                         ) !=
                         null
@@ -263,11 +262,11 @@ class _PostListState extends State<PostList> {
                 return ListTile(
                   isThreeLine: true,
                   leading: GestureDetector(
-                    onTap: onTap,
+                    onTap: onTwterTap,
                     child: Avatar(imageUrl: twt.twter.avatar.toString()),
                   ),
                   title: GestureDetector(
-                    onTap: onTap,
+                    onTap: onTwterTap,
                     child: Text(
                       twt.twter.nick,
                       style: Theme.of(context).textTheme.headline6,
@@ -293,61 +292,66 @@ class _PostListState extends State<PostList> {
                                 );
                               }
 
-                              final image = GestureDetector(
-                                onTap: () async {
-                                  if (isVideoThumbnail) {
-                                    Navigator.push(
-                                      context,
-                                      MaterialPageRoute(
-                                        builder: (context) {
-                                          return VideoScreen(
-                                            title: title ?? '',
-                                            videoURL: uri.toString(),
-                                          );
-                                        },
-                                      ),
-                                    );
-                                    return;
-                                  }
-
-                                  if (await canLaunch(uri.toString())) {
-                                    await launch(uri.toString());
-                                    return;
-                                  }
-
-                                  Scaffold.of(context).showSnackBar(
-                                    SnackBar(
-                                      content: Text('Failed to launch image'),
+                              final onTap = () async {
+                                if (isVideoThumbnail) {
+                                  Navigator.push(
+                                    context,
+                                    MaterialPageRoute(
+                                      builder: (context) {
+                                        return VideoScreen(
+                                          title: title ?? '',
+                                          videoURL: uri.toString(),
+                                        );
+                                      },
                                     ),
                                   );
+                                  return;
+                                }
+
+                                if (await canLaunch(uri.toString())) {
+                                  await launch(uri.toString());
+                                  return;
+                                }
+
+                                Scaffold.of(context).showSnackBar(
+                                  SnackBar(
+                                    content: Text('Failed to launch image'),
+                                  ),
+                                );
+                              };
+
+                              final image = CachedNetworkImage(
+                                httpHeaders: {
+                                  HttpHeaders.acceptHeader: "image/webp"
                                 },
-                                child: CachedNetworkImage(
-                                  httpHeaders: {
-                                    HttpHeaders.acceptHeader: "image/webp"
-                                  },
-                                  imageUrl: newUri.toString(),
-                                  placeholder: (context, url) =>
-                                      CircularProgressIndicator(),
-                                ),
+                                imageUrl: newUri.toString(),
+                                placeholder: (context, url) =>
+                                    CircularProgressIndicator(),
                               );
 
                               if (isVideoThumbnail) {
-                                return Stack(
-                                  alignment: Alignment.center,
-                                  children: [
-                                    image,
-                                    Center(
-                                      child: Icon(
-                                        Icons.play_arrow,
-                                        color: Colors.white,
-                                        size: 100.0,
+                                return GestureDetector(
+                                  onTap: onTap,
+                                  child: Stack(
+                                    alignment: Alignment.center,
+                                    children: [
+                                      image,
+                                      Center(
+                                        child: Icon(
+                                          Icons.play_arrow,
+                                          color: Colors.white,
+                                          size: 100.0,
+                                        ),
                                       ),
-                                    ),
-                                  ],
+                                    ],
+                                  ),
                                 );
                               }
 
-                              return image;
+                              return GestureDetector(
+                                onTap: onTap,
+                                child: image,
+                              );
                             },
                           ),
                           onTapLink: (link) async {
@@ -405,7 +409,6 @@ class _PostListState extends State<PostList> {
                   ),
                 );
               },
-              // addAutomaticKeepAlives: true,
               childCount: widget.twts.length,
             ),
           ),
@@ -418,99 +421,6 @@ class _PostListState extends State<PostList> {
                 ),
               ),
             )
-        ],
-      ),
-    );
-  }
-}
-
-class _PlayPauseOverlay extends StatelessWidget {
-  const _PlayPauseOverlay({Key key, this.controller}) : super(key: key);
-
-  final VideoPlayerController controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      children: <Widget>[
-        AnimatedSwitcher(
-          duration: Duration(milliseconds: 50),
-          reverseDuration: Duration(milliseconds: 200),
-          child: controller.value.isPlaying
-              ? SizedBox.shrink()
-              : Container(
-                  color: Colors.black26,
-                  child: Center(
-                    child: Icon(
-                      Icons.play_arrow,
-                      color: Colors.white,
-                      size: 100.0,
-                    ),
-                  ),
-                ),
-        ),
-        GestureDetector(
-          onTap: () {
-            controller.value.isPlaying ? controller.pause() : controller.play();
-          },
-        ),
-      ],
-    );
-  }
-}
-
-class TwtAssetVideo extends StatefulWidget {
-  final String videoURL;
-
-  const TwtAssetVideo({Key key, this.videoURL}) : super(key: key);
-
-  @override
-  _TwtAssetVideoState createState() => _TwtAssetVideoState();
-}
-
-class _TwtAssetVideoState extends State<TwtAssetVideo> {
-  VideoPlayerController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = VideoPlayerController.network(widget.videoURL);
-
-    _controller.addListener(() {
-      setState(() {});
-    });
-    _controller.setLooping(true);
-    _controller.initialize().then((_) => setState(() {}));
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      child: Column(
-        children: <Widget>[
-          Container(
-            padding: const EdgeInsets.only(top: 20.0),
-          ),
-          Container(
-            padding: const EdgeInsets.all(20),
-            child: AspectRatio(
-              aspectRatio: _controller.value.aspectRatio,
-              child: Stack(
-                alignment: Alignment.bottomCenter,
-                children: <Widget>[
-                  VideoPlayer(_controller),
-                  _PlayPauseOverlay(controller: _controller),
-                  VideoProgressIndicator(_controller, allowScrubbing: true),
-                ],
-              ),
-            ),
-          ),
         ],
       ),
     );
